+++ packages/scrypt-ts-opcat/src/smart-contract/builtin-libs/backtrace.ts
@@ -43,7 +43,7 @@ export class Backtrace extends SmartContractLib {
    * @see packages/scrypt-ts-opcat/test/local-test/genesis.test.ts - GENESIS_SCRIPT_HASH validation tests
    */
   @prop()
-  static readonly GENESIS_SCRIPT_HASH: ByteString = toByteString('8f6157664f8f7cd43e6bba7f3209bb803d4d0fb2fb6e59fdc548989ed3901f15');
+  static readonly GENESIS_SCRIPT_HASH: ByteString = toByteString('836c4b45ab7a625549ca7c799b02d3304adf996557929ee809613e602d26c1f8');
 
   /**
    * Verifies that the transaction hash preimage matches the previous transaction hash